Move bins to a single CLI
