+++ .env
@@ -1,6 +1,6 @@
-DATABASE_URL=postgres://openadr:openadr@localhost:5432/openadr
+DATABASE_URL=postgres://openadr:openadr@localhost:5433/openadr
 VTN_PORT=3000
-PG_PORT=5432
+PG_PORT=5433
 PG_USER=openadr
 PG_DB=openadr
 PG_PASSWORD=openadr
