+++ Cargo.toml
@@ -3,7 +3,7 @@ members = [
     "openleadr-vtn",
     "openleadr-client",
     "openleadr-wire"
-]
+, "cli"]
 exclude = []
 
 resolver = "2"
@@ -21,7 +21,6 @@ description = "An OpenADR 3.0 VTN/VEN implementation"
 [workspace.dependencies]
 openleadr-wire = { version = "0.0.1", path = "openleadr-wire" }
 openleadr-vtn = { version = "0.0.1", path = "openleadr-vtn" }
-openleadr-client = { version = "0.0.1", path = "openleadr-client" }
 
 serde = { version = "1.0.203", features = ["derive"] }
 serde_json = "1.0.117"
@@ -35,7 +34,6 @@ tower = { version = "0.5", features = ["util"] }
 
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
-tracing-test = "0.2.5"
 
 chrono = "0.4.38"
 iso8601-duration = { version = "0.2.0", features = ["chrono"] }
@@ -52,7 +50,6 @@ http-body-util = "0.1.0"
 jsonwebtoken = { version = "9.3.0", default-features = false }
 base64 = "0.22.1"
 rand = "0.8.5"
-async-trait = "0.1.81"
 
 quickcheck = "1.0.3"
