+++ Dockerfile
@@ -8,8 +8,8 @@ WORKDIR /app
 COPY . .
 
 # Don't depend on live sqlx during build use cached .sqlx
-RUN SQLX_OFFLINE=true cargo build --release --bin openleadr-vtn
-RUN cp /app/target/release/openleadr-vtn /app/openleadr-vtn
+RUN SQLX_OFFLINE=true cargo build --release
+RUN cp /app/target/release/cli /app/cli
 
 FROM debian:bookworm-slim as final
 RUN apt-get update && apt-get install curl -y
@@ -22,9 +22,9 @@ ARG gid=2000
 RUN addgroup --gid ${gid} ${group} && adduser --uid ${uid} --gid ${gid} --system --disabled-login --disabled-password ${user}
 EXPOSE 3000
 # get the pre-built binary from builder so that we don't have to re-build every time
-COPY --from=1 --chown=nonroot:nonroot /app/openleadr-vtn/openleadr-vtn /home/nonroot/openleadr-vtn
-RUN chmod 777 /home/nonroot/openleadr-vtn
+COPY --from=1 --chown=nonroot:nonroot /app/cli/cli /home/nonroot/cli
+RUN chmod 777 /home/nonroot/cli
 
 USER $user
 
-ENTRYPOINT ["./home/nonroot/openleadr-vtn"]
+ENTRYPOINT ["./home/nonroot/cli", "vtn"]
