+++ cli/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "cli"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+publish.workspace = true
+description.workspace = true
+
+[dependencies]
+openleadr-vtn = { version = "0.0.1", path = "../openleadr-vtn" }
+openleadr-client = { version = "0.0.1", path = "../openleadr-client" }
+tokio = { workspace = true, features = ["full"] }
+dotenvy.workspace = true
+clap = { version = "4.5.20", features = ["derive"] }
